refactor: 패키지 및 파일 분리
- 상수 분리
- 모델 분리
- utils 분리
