+++ src/main/kotlin/lotto/constants/ErrorMessage.kt
@@ -1,0 +1,4 @@
+package lotto.constants
+
+object ErrorMessage {
+}
