+++ src/main/kotlin/lotto/constants/Message.kt
@@ -1,0 +1,19 @@
+package lotto.constants
+
+object InputMessage {
+    const val PRICE = "구입금액을 입력해 주세요."
+    const val WINNING_NUMBER = "당첨 번호를 입력해 주세요."
+    const val BONUS_NUMBER = "보너스 번호를 입력해 주세요."
+}
+
+object OutputMessage {
+    const val LOTTO_COUNT = "개를 구매했습니다."
+    const val RESULT_TITLE = "당첨 통계\n---"
+    const val MATCH_COUNT = "개 일치"
+    const val MATCH_BONUS_NUMBER = ", 보너스 볼 일치"
+    const val PRICE_OPEN_BRACKET = " ("
+    const val PRICE_CLOSE_BRACKET = "원) - "
+    const val COUNT = "개"
+    const val PROFIT_RATE_START = "총 수익률은 "
+    const val PROFIT_RATE_END = "%입니다."
+}
