+++ src/main/kotlin/lotto/utils/Formatter.kt
@@ -1,0 +1,10 @@
+package lotto.utils
+
+import java.text.DecimalFormat
+
+object Formatter {
+    fun decimalFormatter(number: Int): String {
+        val decimalFormat = DecimalFormat("#,###")
+        return decimalFormat.format(number)
+    }
+}
