+++ src/main/kotlin/lotto/utils/Validation.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.utils
 
 object Validation {
     fun isValidRange(vararg numbers: Int): Boolean = numbers.map { it in 1..45 }.all { it }
